+++ test/unit-disabled/main.tf
@@ -1,0 +1,48 @@
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# EMPTY FEATURES (DISABLED) UNIT TEST
+# This module tests an empty set of features.
+# The purpose is to verify no resources are created when the module is disabled.
+# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+variable "gcp_region" {
+  type        = string
+  description = "(Required) The gcp region in which all resources will be created."
+}
+
+variable "gcp_project" {
+  type        = string
+  description = "(Required) The ID of the project in which the resource belongs."
+}
+
+terraform {
+  required_providers {
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = "~> 4.0"
+    }
+  }
+}
+
+provider "google-beta" {
+  region  = var.gcp_region
+  project = var.gcp_project
+}
+
+# DO NOT RENAME MODULE NAME
+module "test" {
+  source = "../.."
+
+  module_enabled = false
+
+  # add all required arguments
+
+  # add only required arguments and no optional arguments
+  repository_id = "unit-disabled"
+  location      = var.gcp_region
+}
+
+# outputs generate non-idempotent terraform plans so we disable them for now unless we need them.
+# output "all" {
+#   description = "All outputs of the module."
+#   value       = module.test
+# }