+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3]
+
+### Added
+
+- Add unit tests
+
 ### Changed
 
 - Refactored `timeouts` block to match template standard
+- Upgrade [terraform-google-artifact-registry-repository-iam](https://github.com/mineiros-io/terraform-google-artifact-registry-repository-iam) to `v0.0.3`
+- BREAKING: `var.location` needs to be mandatory.
 
 ### Removed
 
@@ -27,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Implementation
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.2...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/mineiros-io/terraform-google-artifact-registry-repository/releases/tag/v0.0.1